services/core/code/pyproject.toml
[tool.poetry]
name = "kaleo-core"
version = "0.1.0"
description = "Kaleo Core"
authors = ["Adrián Pérez <adrianpga.dev@gmail.com>"]
packages = [
    {include = "kaleo_core"},
    {include = "kaleo_core_api_server"}
]

[tool.poetry.dependencies]
python = "^3.11.5"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
pgvector = "^0.4.1"
asyncpg = "^0.30.0"
pyyaml = "^6.0.2"
httpx = "^0.26.0"
python-dotenv = "^1.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.0.1"

[tool.poetry.scripts]
app-start = "kaleo_core.scripts.serve:main"
app-start-lt = "kaleo_core.scripts.serve_lt:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
